+++ core/src/search/LabelIndex.spec.ts
@@ -101,4 +101,68 @@ describe("label index", () => {
       expect(result).toEqual([]);
     });
   });
+
+  describe("contains", () => {
+    it("returns true if the store contains a label for the given uri", () => {
+      const store = graph();
+      store.add(
+        sym("https://pod.example/item#it"),
+        sym("http://www.w3.org/2000/01/rdf-schema#label"),
+        lit("An item"),
+        sym("https://pod.example/label-index"),
+      );
+      const index = new LabelIndex(
+        "https://pod.example/label-index",
+        store,
+        false,
+      );
+      const result = index.contains("https://pod.example/item#it");
+      expect(result).toBe(true);
+    });
+
+    it("returns false if the store does not contain a label for the given uri", () => {
+      const store = graph();
+      const index = new LabelIndex(
+        "https://pod.example/label-index",
+        store,
+        false,
+      );
+      const result = index.contains("https://pod.example/item#it");
+      expect(result).toBe(false);
+    });
+
+    it("returns false if the store contains a label for a different uri", () => {
+      const store = graph();
+      store.add(
+        sym("https://pod.example/other-item#it"),
+        sym("http://www.w3.org/2000/01/rdf-schema#label"),
+        lit("Another item"),
+        sym("https://pod.example/label-index"),
+      );
+      const index = new LabelIndex(
+        "https://pod.example/label-index",
+        store,
+        false,
+      );
+      const result = index.contains("https://pod.example/item#it");
+      expect(result).toBe(false);
+    });
+
+    it("returns false if the store contains a label for the given uri but from a different source", () => {
+      const store = graph();
+      store.add(
+        sym("https://pod.example/item#it"),
+        sym("http://www.w3.org/2000/01/rdf-schema#label"),
+        lit("Label from other source"),
+        sym("https://pod.example/other-source"),
+      );
+      const index = new LabelIndex(
+        "https://pod.example/label-index",
+        store,
+        false,
+      );
+      const result = index.contains("https://pod.example/item#it");
+      expect(result).toBe(false);
+    });
+  });
 });